reform-boot-config: allow messing up the currently running system
Add an interactive prompt that asks the user if they are sure that they
want to rewrite the /boot partition of the currently running system to
boot something else. While this will make the current system unbootable
(and that's why the safeguard was introduced) making the current system
unbootable is desired if the user boots of an SD-Card flashed with the
sysimage and just wants to set up NVMe or eMMC and then use the SD-Card
for /boot. It is thus undesirable to forbid reform-boot-config from
making the current system unbootable as the user might be in the process
of setting up a new system.