etc/skel/.local/bin/reform-windowmaker: Bump $MAX
MAX needs to be increased as 10 is not long enough for xdpyinfo to
initialise things. This results in the first time the script is run
things not working correctly and the user is presented with a thatched
grey screen and X cursor.
Killing Xwayland and restarting then loads Window Maker.
The value in the original gist is 120, but 15 worked for me every time 
on first attempt, following a reboot.