kernel/zz-reform-tools: always print something to STDERR
linux-run-hooks from src:linux-base runs run-parts with --report which
only prints the name of the scripts it runs *if* the script in question
produces any output. Scripts that do not produce any output do not get
an indication by run-parts that they got run. So we print early to make
sure that we are able to easily see if this script was run or not during
a given installation.